+++ Backend/DataCollector/Models/ApplicationRecord.cs
@@ -13,7 +13,7 @@ public class ApplicationRecord
     public int? PositionY { get; set; }
     public int? Width { get; set; }
     public int? Height { get; set; }
-    public int? WindowId { get; set; }
+    public long? WindowId { get; set; }
     
     public ApplicationDto ToDto()
     {
